MOTOR TEMPERATURE SENSOR LIBRARY CONSTRUCTION

+++ STM32CubeIDE/Application/Hardware/motor_param.c
@@ -88,7 +88,7 @@ void calcDC()
 	/*Apply Moving average filtering (Optional)*/
 }
 
-uint16_t getDC()
+int32_t getDC()
 {
-	return ptrMotor.milliAmpere;
+	return (int32_t)ptrMotor.milliAmpere;
 }
