+++ STM32CubeIDE/Application/Hardware/motor_param.h
@@ -55,7 +55,7 @@ void getIqIdMotor();
 
 void calcDC();
 
-uint16_t getDC();
+int32_t getDC();
 
 #ifdef __cplusplus
 }
